Churn Bank Customer Analysis
Project Title: Churn Bank Customer Analysis

I completed this project as part of my final assignment at RevoU, and I am proud to share that our team achieved a top-five ranking for the best project. In this project, we conducted an in-depth analysis to identify the reasons behind customer attrition and to provide business recommendations based on predefined customer clusters.